// src/types.ts
//interface for characters, including different fields that weren't in example
export interface Character{
    id: number;
    name: string;
    image: string;
    species: string;
    status: string;
    gender: string; //new field
    created: string; //new field
    url: string; //new field
    type: string; //new field
}